requirements.txt
requests~=2.32.3
boto3~=1.37.34
dotenv~=0.9.9
python-dotenv~=1.1.0
PyYAML~=6.0.2